+++ CHANGELOG.md
@@ -3,7 +3,9 @@
  - Added q14 (Custom Files)
  - Added Password Generation on Cover Sheet
  - bugfixes & style improvements
-
+ - Added Debugging for SMTP Connection
+ 
+>>>>>>> Stashed changes
 # Version 1.4.0
  - Added Reminder Email to be sent out after 7 days of inactivity with still pending reviews.
  - Added Revision Number on Output Documents